Correct service instantiation in integration tests
